refactor: 쿼리/커멘드 서비스 분리
