+++ src/main/kotlin/com/celuveat/restaurant/application/RestaurantService.kt
@@ -1,0 +1,32 @@
+package com.celuveat.restaurant.application
+
+import com.celuveat.restaurant.application.port.`in`.AddInterestedRestaurantsUseCase
+import com.celuveat.restaurant.application.port.`in`.DeleteInterestedRestaurantsUseCase
+import com.celuveat.restaurant.application.port.`in`.command.AddInterestedRestaurantCommand
+import com.celuveat.restaurant.application.port.`in`.command.DeleteInterestedRestaurantCommand
+import com.celuveat.restaurant.application.port.out.DeleteRestaurantPort
+import com.celuveat.restaurant.application.port.out.SaveRestaurantPort
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+
+@Service
+class RestaurantService(
+    private val saveRestaurantPort: SaveRestaurantPort,
+    private val deleteRestaurantPort: DeleteRestaurantPort,
+) : AddInterestedRestaurantsUseCase, DeleteInterestedRestaurantsUseCase {
+    @Transactional
+    override fun addInterestedRestaurant(command: AddInterestedRestaurantCommand) {
+        saveRestaurantPort.saveInterestedRestaurant(
+            memberId = command.memberId,
+            restaurantId = command.restaurantId,
+        )
+    }
+
+    @Transactional
+    override fun deleteInterestedRestaurant(command: DeleteInterestedRestaurantCommand) {
+        deleteRestaurantPort.deleteInterestedRestaurant(
+            memberId = command.memberId,
+            restaurantId = command.restaurantId,
+        )
+    }
+}
