+++ src/test/kotlin/com/celuveat/restaurant/application/RestaurantQueryServiceTest.kt
@@ -4,12 +4,8 @@ import com.celuveat.celeb.application.port.out.FindCelebritiesPort
 import com.celuveat.celeb.domain.Celebrity
 import com.celuveat.celeb.domain.YoutubeContent
 import com.celuveat.common.application.port.`in`.result.SliceResult
-import com.celuveat.restaurant.application.port.`in`.command.AddInterestedRestaurantCommand
-import com.celuveat.restaurant.application.port.`in`.command.DeleteInterestedRestaurantCommand
 import com.celuveat.restaurant.application.port.`in`.query.GetInterestedRestaurantsQuery
-import com.celuveat.restaurant.application.port.out.DeleteRestaurantPort
 import com.celuveat.restaurant.application.port.out.FindRestaurantPort
-import com.celuveat.restaurant.application.port.out.SaveRestaurantPort
 import com.celuveat.restaurant.domain.Restaurant
 import com.celuveat.support.channelIdSpec
 import com.celuveat.support.sut
@@ -19,20 +15,14 @@ import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.verify
 
-class RestaurantsServiceTest : BehaviorSpec({
-
+class RestaurantQueryServiceTest : BehaviorSpec({
     val findRestaurantPort: FindRestaurantPort = mockk()
     val findCelebritiesPort: FindCelebritiesPort = mockk()
-    val saveRestaurantPort: SaveRestaurantPort = mockk()
-    val deleteRestaurantPort: DeleteRestaurantPort = mockk()
 
-    val restaurantsService = RestaurantsService(
+    val restaurantQueryService = RestaurantQueryService(
         findRestaurantPort,
         findCelebritiesPort,
-        saveRestaurantPort,
-        deleteRestaurantPort,
     )
 
     Given("관심 음식점을 조회할 때") {
@@ -63,42 +53,12 @@ class RestaurantsServiceTest : BehaviorSpec({
                 page = page,
                 size = size,
             )
-            val interestedRestaurants = restaurantsService.getInterestedRestaurant(getInterestedRestaurantsQuery)
+            val interestedRestaurants = restaurantQueryService.getInterestedRestaurant(getInterestedRestaurantsQuery)
 
             Then("관심 식당 목록을 반환한다") {
                 interestedRestaurants.contents.size shouldBe 3
                 interestedRestaurants.contents[0].visitedCelebrities.size shouldBe 2
                 interestedRestaurants.hasNext shouldBe false
-            }
-        }
-    }
-
-    Given("회원이 관심 음식점 추가 시") {
-        val memberId = 1L
-        val restaurantId = 1L
-
-        When("해당 음식점이") {
-            every { saveRestaurantPort.saveInterestedRestaurant(memberId, restaurantId) } returns Unit
-
-            val command = AddInterestedRestaurantCommand(memberId, restaurantId)
-            restaurantsService.addInterestedRestaurant(command)
-            Then("관심 음식점으로 추가 된다") {
-                verify { saveRestaurantPort.saveInterestedRestaurant(memberId, restaurantId) }
-            }
-        }
-    }
-
-    Given("회원이 관심 음식점 삭제 시") {
-        val memberId = 1L
-        val restaurantId = 1L
-
-        When("해당 음식점이") {
-            every { deleteRestaurantPort.deleteInterestedRestaurant(memberId, restaurantId) } returns Unit
-
-            val command = DeleteInterestedRestaurantCommand(memberId, restaurantId)
-            restaurantsService.deleteInterestedRestaurant(command)
-            Then("관심 음식점에서 삭제 된다") {
-                verify { deleteRestaurantPort.deleteInterestedRestaurant(memberId, restaurantId) }
             }
         }
     }
